added android app

+++ nextChallengeAngularWeb/src/app/profile/about/education/education.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppService } from "../../.././services/app.service";
+import { NotificationsService } from "../../.././services/notifications.service";
 import { FormControl, FormGroup, FormArray, FormBuilder, NgForm } from '@angular/forms';
 
 @Component({
@@ -12,7 +13,7 @@ export class EducationComponent implements OnInit {
   public form: FormGroup;
   public UserData: any;
   public SchoolData: any;
-  constructor(private fb: FormBuilder, private _appService: AppService) {
+  constructor(private fb: FormBuilder, private _appService: AppService, private _notificationsService: NotificationsService) {
     this.form = this.fb.group({
       schools: this.fb.array([]),
     });
@@ -20,28 +21,31 @@ export class EducationComponent implements OnInit {
 
   ngOnInit(): void {
     this.UserData = this._appService.getUserData();
-    this._appService.retrieveschools(this.UserData["_id"]).subscribe(data => {
-      const scrools = this.form.controls.schools as FormArray;
-      this.SchoolData = data;
-      this.SchoolData.forEach(element => {
-        scrools.push(this.fb.group({
-          _id: element._id,
-          UserID: element.UserID,
-          Name: element.Name,
-          From: element.From,
-          To: element.To,
-          Description: element.Description
-        }));
+    if (this.UserData != null) {
+      this._appService.retrieveschools(this.UserData["_id"]).subscribe(data => {
+        const scrools = this.form.controls.schools as FormArray;
+        this.SchoolData = data;
+        this.SchoolData.forEach(element => {
+          scrools.push(this.fb.group({
+            _id: element._id,
+            UserID: element.UserID,
+            Name: element.Name,
+            From: element.From,
+            To: element.To,
+            Description: element.Description
+          }));
+        });
       });
-
-    });
+      this._notificationsService.updateChatStatus();
+    }
   }
   sumbitEducation() {
     let data = [];
     this.form.value["schools"].forEach(element => {
       data.push(element);
     });
-    this._appService.updateschools(JSON.stringify(data)).subscribe(data => {});
+    this._appService.updateschools(JSON.stringify(data)).subscribe(data => { });
+    this._notificationsService.updateChatStatus();
   }
   addSchools() {
     const scrools = this.form.controls.schools as FormArray;
@@ -53,13 +57,16 @@ export class EducationComponent implements OnInit {
       To: '',
       Description: ''
     }));
+    this._notificationsService.updateChatStatus();
   }
   cmpare(index) {
     return index;
   }
   deleteEduction(i) {
     const schools = this.form.controls.schools as FormArray;
-    this._appService.deleteschool(schools.at(i).value._id).subscribe(data => {});
+    if (schools.at(i).value._id != null)
+      this._appService.deleteschool(schools.at(i).value._id).subscribe(data => { });
     schools.removeAt(i);
+    this._notificationsService.updateChatStatus();
   }
 }
